requirements.txt
requests==2.31.0
python-dotenv==1.0.0
psycopg==3.1.9
psycopg-binary==3.1.9
websockets==11.0.3
isort>=5.12.0
black>=23.3.0
flake8>=6.0.0
selenium==4.10.0
fredapi==0.5.1
pandas == 2.0.1
numpy == 1.24.3
tqdm == 4.66.1